enh: move socket to folder, client
