+++ client/src/services/socket.js
@@ -1,10 +1,10 @@
 import io from "socket.io-client";
-import store from "./store";
+import store from "../store";
 import {
   setNewMessage,
   removeOfflineUser,
   addOnlineUser,
-} from "./store/conversations";
+} from "../store/conversations";
 
 const socket = io(window.location.origin);
